I would like you to implememt the first file of code. Create a pkg/yaml package. Then a pkg/yaml/extract.go file. There will be a ExtractCommon() function there. This function will extract the common components in two YAML files, returning
1) the common structure
2) the two YAMLs without that common structure.

For example, for these two YAML file (lets call them YAML-1 and YAML-2):

```yaml
foo:
  bar:
    something: [1,2,3]
    other: true
```

and

```yaml
foo:
  bar:
    else: [1,2,3]
    other: true
```

the common structure (lets call it YAML-COMMON-1) would be:

```yaml
foo:
  bar:
    other: true
```

and the two YAMLs without the common part (lets call them UPDATED-YAML-1 and UPDATED-YAML-2) would be

```yaml
foo:
  bar:
    something: [1,2,3]
```

and

```yaml
foo:
  bar:
    else: [1,2,3]
```

It is important to note that merging the common YAML returns the original YAMLs, so:

UPDATED-YAML-1 + YAML-COMMON-1 = YAML-1
UPDATED-YAML-2 + YAML-COMMON-1 = YAML-2

Please add some options in the ExtractCommon() function, so we can customize the whole
process with some WithFeature() arguments.

Take into account that, if a list has the same elements in YAML-1 and YAML-2, we should consider
it to be part of the YAML-COMMON-1. But please add a flag for not following this behaviour.

Add extensive unit tests for these operations. Use fixtures, or YAML files embeded
in the tests files, so it is easy to read an understand by a human what is being tested,
as well as the expected output. Make sure unit tests pass. If not, debug and fix the problem.

Signed-off-by: Alvaro Saurin <alvaro.saurin@gmail.com>