I have found a corner case where the ExtractCommonRecursive logic will fail to find a common subset of the YAML.
Imagine we have /a/1/X/values.yaml with;

```yaml
foo:
  bar:
    something: [1,2,3]
    other: true
```

and /a/2/Y/values.yaml with;

```yaml
foo:
  bar:
    else: [1,2,3]
    other: true
```

They are very similar, but they are sibilings at at deeper lever (ie, they have common grandparents).
So we should create the common YAML file at /a/common.yaml.
Could we solve this is an easy way? Does it make sense?
If that is the case, please implement it.
Add exhaustive unit tests, and make sure they pass.

Signed-off-by: Alvaro Saurin <alvaro.saurin@gmail.com>