I have found another corner case. I think we should try to find a generalized solution.
Imagine we have /a/1/X/values.yaml with;

```yaml
foo:
  bar:
    something: [1,2,3]
    other: true
```

and /a/2/values.yaml with;

```yaml
foo:
  bar:
    else: [1,2,3]
    other: true
```

They are very similar, but they are not even sibilings.
But we we should create the common YAML file at /a/common.yaml.
Could we deal with this case? If that is the case, please implement it.
Add exhaustive unit tests, and make sure they pass.

Signed-off-by: Alvaro Saurin <alvaro.saurin@gmail.com>